+++ src/public_api.ts
@@ -4,5 +4,6 @@
 export * from './lib/avatar-config';
 export * from './lib/avatar.component';
 export * from './lib/avatar.module';
+export * from './lib/avatar.providers';
 export * from './lib/avatar.service';
 export * from './lib/sources/avatar-source.enum';